requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
flake8==6.0.0
httpie==3.2.1
idna==3.4
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.1
requests-toolbelt==0.10.1
rich==13.0.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
